backend/requirements.txt
backend==0.2.4.1
cryptography==45.0.4
fastapi==0.115.14
firebase_admin==6.8.0
google_auth_oauthlib==1.2.2
httpx==0.28.1
iron_cache==0.3.2
langchain==0.3.26
langchain_community==0.3.26
langchain_core==0.3.66
langchain_groq==0.3.4
langchain_huggingface==0.3.0
langchain_pinecone==0.2.8
langsmith==0.4.3
nltk==3.9.1
pdf2image==1.17.0
pdfplumber==0.11.6
Pillow==11.2.1
pinecone==7.2.0
pydantic==2.11.7
pyotp==2.9.0
pytesseract==0.3.13
pytest==8.4.1
python-dotenv==1.1.1
python_docx==1.2.0
qrcode==8.2
reportlab==4.4.2
sentence_transformers==4.1.0
streamlit==1.44.1
supabase==2.16.0
torch==2.4.0
urllib3==2.5.0
uvicorn==0.34.3
